LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/referrer-policies.sub-expected.txt

CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/referrer-checker.py?name=remoteNoReferrerPolicy
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/referrer-checker.py?name=remoteReferrerPolicyOrigin
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/referrer-checker.py?name=remoteReferrerPolicyNoReferrer
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/json-module/referrer-checker.py?name=remoteNoReferrerPolicyUnsafeUrl

Harness Error (TIMEOUT), message = null
